Update food detail and review 2

+++ app/src/main/java/com/example/dormdeli/model/Review.kt
@@ -7,7 +7,7 @@ data class Review(
     val userAvatarUrl: String = "",
     val storeId: String = "",
     val foodId: String = "",
-    val rating: Long = 0,
+    val rating: Int = 0,
     val comment: String = "",
     val createdAt: Long = System.currentTimeMillis()
 )
